fixup! fix(ivy): move HostListeners generation to factory function
